requirement.txt
asgiref==3.7.2
Cython==3.0.0
Django==3.2.19
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
inflection==0.5.1
mysqlclient==2.2.0
packaging==23.1
Pillow==8.0.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==5.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1